Move validate-project-path.ts to helpers/
